serv_1/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
databases
psycopg2-binary
asyncpg
pika
boto3
requests